parse command-line into a central Options struct
